scrolling through full images works (with code rewrite)
Lots of changes and rewriting of how the code works. Because all the
changes were happening simultaneously I couldn't commit it in smaller
batches. Lesson for next time: one problem at a time.

-Added fullLoaderCache that holds a reference to the Loader of the full
image while the image is still not loaded. As soon as it's loaded,
fullLoaded() is called and the loaded image is assigned to fullLoader.
Adding fullLoaderCache fixed a bug in which scrolling through full
images wouldn't work because fullLoader was being manipulated before
the image was loaded. This meant that parts of the code were rewritten
to work in fullLoaderCache.
-Added loadFull() to reduce repetitive code
-Changed fullLoaded() to not add the fullLoader to the screen right
away and then turn it invisible while waiting for its animation.
Instead, fullLoaded only adds the loaded image to fullLoader and it's
added to the stage once it's time to animate the full image in (in
animateFullInAfterGrid()).
-Added functionality to the prev and next buttons: they now scroll
through the full images. An event listener is placed on the buttons
once they are added to the stage (and also removed once they are
removed from the stage, this is done ine showHideButtons()) that starts
animating the image away (goToNext() and goToPrev()). Once that tween
is finished, it calls a new function that tweens in the following image
(loadNext() and loadPrev()).
-The event listener on the thumbs that loads in the full image is now
only called once all the thumbs are animated in. Therefore I removed
two instances of the event listener being added and now I only add it
at the end of the grid animation through makeGridActive()
-Removed loader:Loader parameters and hard coded fullLoader into most
of the places. Not optimal and easily usable in other code, but because
of the deadline it is necessary.
-Changed lastThumbAnimatedAway() name into makeGridActive() to better
reflect its function. Also changed how it works: instead of being a
function that returns a function (to make it be able to be called
through event listeners while still being able to get parameters) it
now returns void and doesn't need parameters as it directly interacts
with fullLoader.
-Added comments to explain all of the arguments used in animateTween
-Changed for-loop in animateGridIn() to call makeGridActive() when last
thumb is animated in.
-Moved shuffleArray to the end.

-Lots of comments