language button added, everything working as intended
-Added langBtn and the necessary code to make the button work
(assigning, event listeners, adding it to showHideButtons())
-Added changeLanguage() which changes the language boolean (nl) but had
most of the code of removeFull().
-Added animateFullOut() which does what was removeFull() (and
changeLanguage()) with an added if-else to call the correct function
with its event listener. Parameter decides which function is called:
whether grid is loaded in, or image is reloaded in other language.
-removeFull() is simplified along with changeLanguage()
-changed animateFullInAfterGrid() to
animateFullIn() because it's now used more generally. Also calling
loadFull() from here now.