tween-stop bug fixed
Turns out the tweens were garbage collected at random intervals because
they weren't stored until they finished. That is fixed now. The program
works the way it's supposed to.

-Added tweenList to hold the tweens while they are playing.
-addToTweenList() is a function that passes the supplied tween to the
tweenList and adds an event listener for the end of the tween to call
removeFromTweenList() to remove the finished tween from the list (and
be eligible for garbage collection again).
-Added addToTweenList() after every instance of a tween being created.
-Moved animateFullOut() and makeGridActive() around.
-Changed the visuals of the buttons and had to change the starting
position of the buttons. Moved it left by 27 pixels.